+++ benchmark_test.go
@@ -4,27 +4,47 @@ import (
 	"testing"
 	"time"
 
-	. "github.com/gbrlsnchs/jwt"
+	. "github.com/gbrlsnchs/jwt/v2"
 )
 
 const benchMock = "eyJhbGciOiJIUzI1NiIsImtpZCI6ImtpZCIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE1MzU5NDE1NjcsImV4cCI6MTU2NzA0NTU2NywibmJmIjoxNTM1OTQzMzY3LCJqdGkiOiJCZW5jaG1hcmtTaWduIiwiYXVkIjoiYmVuY2htYXJrIiwic3ViIjoibWUiLCJpc3MiOiJnYnJsc25jaHMiLCJuYW1lIjoiZm9vYmFyIiwiaXNCZW5jaCI6dHJ1ZX0.bJSm0om7-BRCuLbICllYEAH7YsAT1cW2fdSfKcMnhOg"
 
 var benchSigner = HS256("benchmark")
 
+type benchToken struct {
+	*JWT
+	Name    string `json:"name,omitempty"`
+	IsBench bool   `json:"isBench"`
+}
+
 func BenchmarkSign(b *testing.B) {
 	now := time.Now()
-	opt := &Options{
-		Timestamp:      true,
-		JWTID:          b.Name(),
-		ExpirationTime: now.Add(24 * 30 * 12 * time.Hour),
-		NotBefore:      now.Add(30 * time.Minute),
-		Subject:        "me",
-		Audience:       "benchmark",
-		Issuer:         "gbrlsnchs",
+	jot := &benchToken{
+		JWT: &JWT{
+			Header: &Header{
+				Algorithm: benchSigner.String(),
+				KeyID:     "kid",
+			},
+			Claims: &Claims{
+				ID:         b.Name(),
+				IssuedAt:   now.Unix(),
+				Expiration: now.Add(24 * 30 * 12 * time.Hour).Unix(),
+				NotBefore:  now.Add(30 * time.Minute).Unix(),
+				Subject:    "me",
+				Audience:   "benchmark",
+				Issuer:     "gbrlsnchs",
+			},
+		},
+		Name:    "foobar",
+		IsBench: true,
 	}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		if _, err := Sign(benchSigner, opt); err != nil {
+		payload, err := Marshal(jot)
+		if err != nil {
+			b.Fatal(err)
+		}
+		if _, err = benchSigner.Sign(payload); err != nil {
 			b.Fatal(err)
 		}
 	}
@@ -32,11 +52,15 @@ func BenchmarkSign(b *testing.B) {
 
 func BenchmarkVerify(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		jot, err := FromString(benchMock)
+		payload, sig, err := Parse(benchMock)
 		if err != nil {
 			b.Fatal(err)
 		}
-		if err = jot.Verify(benchSigner); err != nil {
+		var jot benchToken
+		if err = Unmarshal(payload, &jot); err != nil {
+			b.Fatal(err)
+		}
+		if err = benchSigner.Verify(payload, sig); err != nil {
 			b.Fatal(err)
 		}
 	}