+++ ecdsa.go
@@ -39,11 +39,36 @@ func ES512(priv *ecdsa.PrivateKey, pub *ecdsa.PublicKey) Signer {
 	return &ecdsasha{priv: priv, pub: pub, hash: sha512.New, alg: MethodES512}
 }
 
-func (e *ecdsasha) Sign(msg []byte) ([]byte, error) {
+func (e *ecdsasha) Sign(payload []byte) ([]byte, error) {
 	if e.priv == nil {
 		return nil, ErrNoECDSAPrivKey
 	}
+	sig, err := e.sign(payload)
+	if err != nil {
+		return nil, err
+	}
+	return build(e, payload, sig), nil
+}
 
+func (e *ecdsasha) String() string {
+	return e.alg
+}
+
+func (e *ecdsasha) Verify(payload, sig []byte) (err error) {
+	if e.pub == nil {
+		return ErrNoECDSAPubKey
+	}
+	decSig := make([]byte, enc.DecodedLen(len(sig)))
+	if _, err = enc.Decode(decSig, sig); err != nil {
+		return err
+	}
+	if err = e.verify(payload, decSig); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (e *ecdsasha) sign(msg []byte) ([]byte, error) {
 	hh := e.hash()
 	var err error
 	if _, err = hh.Write(msg); err != nil {
@@ -65,15 +90,7 @@ func (e *ecdsasha) Sign(msg []byte) ([]byte, error) {
 	return append(rsig, ssig...), nil
 }
 
-func (e *ecdsasha) String() string {
-	return e.alg
-}
-
-func (e *ecdsasha) Verify(msg, sig []byte) error {
-	if e.pub == nil {
-		return ErrNoECDSAPubKey
-	}
-
+func (e *ecdsasha) verify(msg, sig []byte) error {
 	byteSize := e.byteSize(e.pub.Params().BitSize)
 	if len(sig) != byteSize*2 {
 		return ErrECDSASigLen